Design - exported each modules to CamBam files.
I have exported each module and the front (except the autopilot
that requires a bit more work because of the Nose Up/Dn) to
separate DXF (excluded) that were then opened in CamBam and saved
to cb files.

The design in the CamBam files might slightly differ because of
fine tuning.

TODO
 - split the front board into 2, so it fits in the laser
cutter machine and... cut !
 - modify and export the autopilot module + top.