Electronic - Added -cache.lib files
According to KiCad documentation, these files should
be included in the project because they ensure that
others can see my symbols (even if they do not have
the external libraries installed).
Even if my external libraries are submodules of this
repository, I prefer to make sure there is no problem.